s4o_requirements.txt
altair==4.2.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.5
bokeh==2.4.3
bqplot==0.12.40
branca==0.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
colour==0.1.5
comm==0.1.4
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.2
Fiona==1.9.4.post1
folium==0.14.0
fonttools==4.42.1
gdown==4.7.1
geojson==3.0.1
geopandas==0.13.2
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.0.0
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.17.3
ipython==8.14.0
ipytree==0.2.2
ipywidgets==8.1.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.17.3
jupyterlab-widgets==3.0.8
kiwisolver==1.4.5
leafmap==0.23.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.39
protobuf==3.20.3
pure-eval==0.2.2
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyproj==3.6.0
pyrsistent==0.19.3
pyshp==2.3.1
PySocks==1.7.1
pystac==1.8.3
pystac-client==0.7.3
python-box==7.1.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rafina==0.0.6
requests==2.28.2
rich==13.3.2
scipy==1.15.1
scooby==0.7.2
semver==2.13.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
stack-data==0.6.2
streamlit==1.40.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
traittypes==0.2.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.15
validators==0.20.0
watchdog==2.3.1
wcwidth==0.2.6
whitebox==2.3.1
whiteboxgui==2.3.0
widgetsnbextension==4.0.8
xyzservices==2023.2.0
zipp==3.15.0
